poprawa

+++ auctions/src/main/java/com/auctions/system/module/auction_processing/dao/AuctionProcessingDAO.java
@@ -7,6 +7,6 @@ public interface AuctionProcessingDAO {
 	
 	public UserDetails getSellerDetails(long userId);
 	
-	public AuctionDetails getAuctionDetails(int auctionId);
+	public AuctionDetails getAuctionDetails(long auctionId);
 
 }
